save layouts order after editing
We update the function ´setEditing´ that is executed after clicking on ´Done Editing´ button to set editing
to false.
We update it to savethe edited layout order, because our UI should be using that new order after  edition.
We pass ´editedlayoutsOrder´ and ´setEditedlayoutsOrder´ to DashboardPage via the custom hook ´useDashboard´
and there we can pass them to´DashboardGrid´ where we need to use them