make sure parameters are disabled during dashboard refresh
We have some queries that take a long time to execute , so the related widgets take a long time to refresh..
This causes a weird behavior because the user may update the parameters while the widgets are still getting refreshed
in this case these widgets at the end will show the old values related to the first change and that is an issue.

Here we disable the possibility to update the parameters while a dashboard is still refreshing.